Conflicts:
	GraphNet/dataset.py
	GraphNet/notebooks/plot_ldmx_nn.ipynb
	GraphNet/run_eval.job
	GraphNet/run_prediction.job
	GraphNet/run_training.job
	GraphNet/train.py

+++ GraphNet/utils/ParticleNet.py
@@ -5,7 +5,7 @@ import torch
 import torch.nn as nn
 
 # NEW:  32-bit float resolution (23 for decimal) isn't enough for PN scores ~ 1 - 1e-6.
-torch.set_default_dtype(torch.float64)
+torch.set_default_dtype(torch.float32)
 
 '''Based on https://github.com/WangYueFt/dgcnn/blob/master/pytorch/model.py.'''
 
@@ -128,7 +128,7 @@ class ParticleNet(nn.Module):
                  **kwargs):
         super(ParticleNet, self).__init__(**kwargs)
 
-        self.bn_fts = nn.BatchNorm1d(input_dims) 
+        self.bn_fts = nn.BatchNorm1d(input_dims)
 
         self.edge_convs = nn.ModuleList()
         for idx, layer_param in enumerate(conv_params):
